Fix critical concurrency issues and implement global resource management
- Add global worker pool to prevent resource explosion
- Implement browser connection pooling for efficiency
- Remove user-controlled concurrency for security
- Fix race conditions with single mutex approach
- Add graceful overload handling with HTTP 503
